FOI-45 Prep work for new form
This does a few things:

- Introduces separate directories to keep the templates for different approaches separate.
  My hope is that this will help with maintainability
- Splits routes.py into three files (one for our one shared route, one for routes specific
  to the multiple forms journey, and one for our single form journey)
- Adds a 'start' page and route for the multiple forms journey, along with an entry in content.yaml
