+++ app/main/routes_single_form_journey.py
@@ -7,13 +7,6 @@ from app.main.forms.request_a_service_record import RequestAServiceRecord
 from flask import redirect, render_template, session, url_for
 
 
-@bp.route("/")
-@cache.cached(key_prefix=cache_key_prefix)
-def index():
-    content = load_content()
-    return render_template("main/index.html", content=content)
-
-
 @bp.route("/all-fields-in-one-form/", methods=["GET", "POST"])
 def all_fields_in_one_form():
     form = RequestAServiceRecord()
@@ -28,7 +21,7 @@ def all_fields_in_one_form():
         return redirect(url_for("main.review"))
 
     return render_template(
-        "main/all-fields-in-one-form.html", content=content, form=form
+        "main/all-fields-in-one-form/form.html", content=content, form=form
     )
 
 
@@ -42,7 +35,7 @@ def review():
         return redirect(url_for("main.send_to_gov_pay"))
 
     return render_template(
-        "main/review.html", form=form, form_data=form_data, content=content
+        "main/all-fields-in-one-form/review.html", form=form, form_data=form_data, content=content
     )
 
 
@@ -80,16 +73,16 @@ def handle_gov_uk_pay_response():
 @bp.route("/payment-link-creation_failed/")
 def payment_link_creation_failed():
     content = load_content()
-    return render_template("main/payment-link-creation-failed.html", content=content)
+    return render_template("main/payment/payment-link-creation-failed.html", content=content)
 
 
 @bp.route("/payment-incomplete/")
 def payment_incomplete():
     content = load_content()
-    return render_template("main/payment-incomplete.html", content=content)
+    return render_template("main/payment/payment-incomplete.html", content=content)
 
 
 @bp.route("/confirm-payment-received/")
 def confirm_payment_received():
     content = load_content()
-    return render_template("main/confirm-payment-received.html", content=content)
+    return render_template("main/payment/confirm-payment-received.html", content=content)
