+++ app/templates/main/all-fields-in-one-form/confirm-payment-received.html
