+++ app/templates/main/multi-page-journey/start.html
@@ -1,0 +1,16 @@
+{% extends 'base.html' %}
+
+{%- set pageTitle = content.app.title -%}
+
+{% block content %}
+  <div class="tna-section">
+    <div class="tna-container">
+      <div
+        class="tna-column tna-column--width-2-3 tna-column--width-5-6-medium tna-column--full-small tna-column--full-tiny">
+
+        <h1 class="tna-heading-xl">{{ content.pages.start.heading }}</h1>
+
+      </div>
+    </div>
+  </div>
+{% endblock %}
