+++ app/templates/main/payment/payment-link-creation-failed.html
